requirements.txt
langchain
langchain-community
langchain-openai
chromadb
python-dotenv
requests
sentence-transformers
faiss-cpu 
fastapi